# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: MESH API
  version: 0.3.0
  license: { name: UNLICENSED, identifier: LicenseRef-UNLICENSED }
  description: |
    Contract for the MESH student-collaboration modular monolith. This milestone exposes account,
    profile, self-declared skills, goals, interests, desired skills, weekly availability, GitHub public-repository evidence,
    discovery, connections, project groups, persistent messages, tasks, calendar events, and availability polls.
    GitHub evidence is a limited supporting signal, never proof of expertise. LinkedIn is not collected or scraped.
servers:
  - url: /api/v1
security:
  - bearerAuth: []
paths:
...
  /profile/me:
    get:
      operationId: getMyProfile
      summary: Get the signed-in student's profile and self-declared skills
      responses:
        '200': { description: Profile, content: { application/json: { schema: { $ref: '#/components/schemas/Profile' } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
    put:
      operationId: updateMyProfile
      summary: Replace editable profile fields for the signed-in student
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/UpdateProfileRequest' } } }
      responses:
        '200': { description: Updated profile, content: { application/json: { schema: { $ref: '#/components/schemas/Profile' } } } }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /profile/me/skills:
    put:
      operationId: replaceMySkills
      summary: Replace all self-declared skills; this does not create evidence or verify expertise
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/ReplaceSkillsRequest' } } }
      responses:
        '200': { description: Updated profile, content: { application/json: { schema: { $ref: '#/components/schemas/Profile' } } } }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /profile/me/goals:
    put:
      operationId: replaceMyGoals
      summary: Replace the signed-in student's collaboration goals
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/ReplaceGoalsRequest' } } }
      responses:
        '200': { description: Updated profile, content: { application/json: { schema: { $ref: '#/components/schemas/Profile' } } } }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /profile/me/interests:
    put:
      operationId: replaceMyInterests
      summary: Replace the signed-in student's interests from the curated catalogue
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/ReplaceInterestsRequest' } } }
      responses:
        '200': { description: Updated profile, content: { application/json: { schema: { $ref: '#/components/schemas/Profile' } } } }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /profile/me/desired-skills:
    put:
      operationId: replaceMyDesiredSkills
      summary: Replace skills the signed-in student is looking for in a collaborator
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/ReplaceDesiredSkillsRequest' } } }
      responses:
        '200': { description: Updated profile, content: { application/json: { schema: { $ref: '#/components/schemas/Profile' } } } }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /profile/me/availability:
    put:
      operationId: replaceMyAvailability
      summary: Replace weekly availability windows for matching and future availability polls
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/ReplaceAvailabilityRequest' } } }
      responses:
        '200': { description: Updated profile, content: { application/json: { schema: { $ref: '#/components/schemas/Profile' } } } }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /skills:
    get:
      operationId: listSkills
      summary: List the curated skill catalogue
      responses:
        '200': { description: Skill catalogue, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/Skill' } } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /goals:
    get:
      operationId: listGoals
      summary: List collaboration goals available for profiles
      responses:
        '200': { description: Goal catalogue, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/Goal' } } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /interests:
    get:
      operationId: listInterests
      summary: List interests available for profiles
      responses:
        '200': { description: Interest catalogue, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/Interest' } } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
...
  /projects:
    post:
      operationId: createProject
      summary: Create a project group with the signed-in student as owner
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/ProjectInput' } } }
      responses:
        '201': { description: Project created, content: { application/json: { schema: { $ref: '#/components/schemas/Project' } } } }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
    get:
      operationId: listMyProjects
      summary: List projects the signed-in student belongs to
      responses:
        '200': { description: Projects, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/Project' } } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /projects/{projectId}:
    get:
      operationId: getProject
      summary: Get one project and its members
      parameters: [ { $ref: '#/components/parameters/ProjectId' } ]
      responses:
        '200': { description: Project, content: { application/json: { schema: { $ref: '#/components/schemas/Project' } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    put:
      operationId: updateProject
      summary: Update a project; owner only
      parameters: [ { $ref: '#/components/parameters/ProjectId' } ]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/ProjectInput' } } }
      responses:
        '200': { description: Updated project, content: { application/json: { schema: { $ref: '#/components/schemas/Project' } } } }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      operationId: deleteProject
      summary: Delete a project; owner only
      parameters: [ { $ref: '#/components/parameters/ProjectId' } ]
      responses:
        '204': { description: Project deleted }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /projects/{projectId}/members:
    post:
      operationId: addProjectMember
      summary: Add one of the owner's established connections to a project
      parameters: [ { $ref: '#/components/parameters/ProjectId' } ]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/AddProjectMemberRequest' } } }
      responses:
        '200': { description: Project with the added member, content: { application/json: { schema: { $ref: '#/components/schemas/Project' } } } }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
  /projects/{projectId}/members/{userId}:
    delete:
      operationId: removeProjectMember
      summary: Remove a non-owner project member; owner only
      parameters: [ { $ref: '#/components/parameters/ProjectId' }, { $ref: '#/components/parameters/UserId' } ]
      responses:
        '204': { description: Member removed }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /projects/{projectId}/ownership:
    patch:
      operationId: transferProjectOwnership
      summary: Transfer project ownership to an existing member; owner only
      parameters: [ { $ref: '#/components/parameters/ProjectId' } ]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/TransferProjectOwnershipRequest' } } }
      responses:
        '200': { description: Project with updated ownership, content: { application/json: { schema: { $ref: '#/components/schemas/Project' } } } }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
  /projects/{projectId}/messages:
    get:
      operationId: listProjectMessages
      summary: List recent persistent messages in chronological order; project members only
      parameters:
        - $ref: '#/components/parameters/ProjectId'
        - name: limit
          in: query
          required: false
          schema: { type: integer, minimum: 1, maximum: 100, default: 50 }
      responses:
        '200': { description: Messages, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/ProjectMessage' } } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    post:
      operationId: sendProjectMessage
      summary: Send a persistent project message; project members only
      parameters: [ { $ref: '#/components/parameters/ProjectId' } ]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/SendProjectMessageRequest' } } }
      responses:
        '201': { description: Message sent, content: { application/json: { schema: { $ref: '#/components/schemas/ProjectMessage' } } } }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
...
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT }
  responses:
    ValidationError:
      description: Request field validation failed
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    Unauthorized:
      description: Missing, invalid, or expired bearer token
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    Conflict:
      description: A unique value is already in use
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    Forbidden:
      description: The signed-in student is not allowed to perform this action
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    NotFound:
      description: The requested resource does not exist or is not accessible
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
...
  parameters:
...
    ProjectId:
      name: projectId
      in: path
      required: true
      schema: { type: string, format: uuid }
    UserId:
      name: userId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  schemas:
...
    UpdateProfileRequest:
      type: object
      required: [displayName]
      properties:
        displayName: { type: string, minLength: 1, maxLength: 100 }
        department: { type: [string, 'null'], maxLength: 120 }
        yearOfStudy: { type: [integer, 'null'], minimum: 1, maximum: 8 }
        bio: { type: [string, 'null'], maxLength: 500 }
        avatarKey: { type: [string, 'null'], maxLength: 60 }
        availability: { type: [string, 'null'], maxLength: 120, description: Optional short human-readable availability summary. }
        primaryDomain: { type: [string, 'null'], maxLength: 80 }
        onboardingComplete: { type: boolean, default: false }
    ReplaceSkillsRequest:
      type: object
      required: [skills]
      properties:
        skills:
          type: array
          maxItems: 50
          items:
            type: object
            required: [skillId, proficiency]
            properties:
              skillId: { type: integer, format: int64 }
              proficiency: { type: integer, minimum: 1, maximum: 5, description: "Student self-assessment, not a verified score." }
    ReplaceGoalsRequest:
      type: object
      required: [goalCodes]
      properties:
        goalCodes: { type: array, maxItems: 7, items: { type: string, example: HACKATHON } }
    ReplaceInterestsRequest:
      type: object
      required: [interestIds]
      properties:
        interestIds: { type: array, maxItems: 12, items: { type: integer, format: int64 } }
    ReplaceDesiredSkillsRequest:
      type: object
      required: [skillIds]
      properties:
        skillIds: { type: array, maxItems: 12, items: { type: integer, format: int64 } }
    ReplaceAvailabilityRequest:
      type: object
      required: [timezone, windows]
      properties:
        timezone: { type: string, description: IANA timezone, example: Asia/Kolkata }
        windows: { type: array, maxItems: 28, items: { $ref: '#/components/schemas/AvailabilityWindow' } }
    Profile:
      type: object
      required: [id, username, displayName, onboardingComplete, skills, goals, interests, desiredSkills, availabilityWindows]
      properties:
        id: { type: string, format: uuid }
        username: { type: string }
        displayName: { type: string }
        department: { type: [string, 'null'] }
        yearOfStudy: { type: [integer, 'null'] }
        bio: { type: [string, 'null'] }
        avatarKey: { type: [string, 'null'] }
        availability: { type: [string, 'null'] }
        availabilityTimezone: { type: [string, 'null'], description: IANA timezone for `availabilityWindows`. }
        primaryDomain: { type: [string, 'null'] }
        onboardingComplete: { type: boolean }
        skills: { type: array, items: { $ref: '#/components/schemas/ProfileSkill' } }
        goals: { type: array, items: { $ref: '#/components/schemas/Goal' } }
        interests: { type: array, items: { $ref: '#/components/schemas/Interest' } }
        desiredSkills: { type: array, items: { $ref: '#/components/schemas/Skill' } }
        availabilityWindows: { type: array, items: { $ref: '#/components/schemas/AvailabilityWindow' } }
    Skill:
      type: object
      required: [id, name, category]
      properties:
        id: { type: integer, format: int64 }
        name: { type: string }
        category: { type: string }
    ProfileSkill:
      allOf:
        - $ref: '#/components/schemas/Skill'
        - type: object
          required: [proficiency, evidenceSupported]
          properties:
            proficiency: { type: integer, minimum: 1, maximum: 5, description: Self-declared proficiency. }
            evidenceSupported: { type: boolean, description: "True only when an authorized GitHub public-repository language matches the self-declared skill; it is a supporting signal, not proof of expertise." }
    Goal:
      type: object
      required: [code, name]
      properties:
        code: { type: string, example: HACKATHON }
        name: { type: string, example: Hackathon project }
    Interest:
      type: object
      required: [id, name]
      properties:
        id: { type: integer, format: int64 }
        name: { type: string }
...
    AvailabilityWindow:
      type: object
      required: [dayOfWeek, startMinute, endMinute]
      properties:
        dayOfWeek: { type: integer, minimum: 1, maximum: 7, description: ISO weekday number; 1 is Monday. }
        startMinute: { type: integer, minimum: 0, maximum: 1439, description: Minutes after midnight in `timezone`. }
        endMinute: { type: integer, minimum: 1, maximum: 1440, description: Must be later than `startMinute`. }
...
    ProjectInput:
      type: object
      required: [name]
      properties:
        name: { type: string, minLength: 1, maxLength: 120, example: Campus accessibility map }
        summary: { type: [string, 'null'], maxLength: 1000, example: Map accessible routes around campus. }
        domain: { type: [string, 'null'], maxLength: 80, example: WEB_DEVELOPMENT }
    AddProjectMemberRequest:
      type: object
      required: [userId]
      properties:
        userId: { type: string, format: uuid }
    TransferProjectOwnershipRequest:
      type: object
      required: [newOwnerUserId]
      properties:
        newOwnerUserId: { type: string, format: uuid }
    Project:
      type: object
      required: [id, name, ownerUserId, createdAt, members]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        summary: { type: [string, 'null'] }
        domain: { type: [string, 'null'] }
        ownerUserId: { type: string, format: uuid }
        createdAt: { type: string, format: date-time }
        members: { type: array, items: { $ref: '#/components/schemas/ProjectMember' } }
    ProjectMember:
      type: object
      required: [userId, username, displayName, role]
      properties:
        userId: { type: string, format: uuid }
        username: { type: string }
        displayName: { type: string }
        avatarKey: { type: [string, 'null'] }
        role: { type: string, enum: [OWNER, MEMBER] }
    SendProjectMessageRequest:
      type: object
      required: [content]
      properties:
        content: { type: string, minLength: 1, maxLength: 2000 }
    ProjectMessage:
      type: object
      required: [id, projectId, senderId, senderName, content, createdAt]
      properties:
        id: { type: string, format: uuid }
        projectId: { type: string, format: uuid }
        senderId: { type: string, format: uuid }
        senderName: { type: string }
        content: { type: string }
        createdAt: { type: string, format: date-time }
...
    Error:
      type: object
      required: [timestamp, status, error, message, fields, path]
      properties:
        timestamp: { type: string, format: date-time }
        status: { type: integer }
        error: { type: string }
        message: { type: string }
        fields: { type: object, additionalProperties: { type: string } }
        path: { type: string }